simplify BuildDataCopierViaMemcpy
